requirements.txt
flask==2.0.3
openai
python-dotenv
flask-sqlalchemy==2.5.1
flask-login==0.5.0
bcrypt==4.0.1
requests==2.31.0
python-dateutil==2.8.2
numpy
pandas
scikit-learn